requirements.txt
# --- requirements.txt ---
pyTelegramBotAPI
google-generativeai
tiktoken